+++ logprep/processor/generic_adder/rule.py
@@ -71,20 +71,6 @@ In the following example two files are being used, but only the first existing f
         - PATH_TO_FILE_WITH_LIST
     description: '...'
 
-It is also possible to use a table from a MySQL database to add fields to an event.
-
-..  code-block:: yaml
-    :linenos:
-    :caption: Example with a MySQL Table
-
-    filter: '*'
-    generic_adder:
-      sql_table:
-        event_source_field: source
-        pattern: '([a-zA-Z0-9]+)_\S+'
-        destination_field_prefix: nested.dict
-    description: '...'
-
 .. autoclass:: logprep.processor.generic_adder.rule.GenericAdderRule.Config
    :members:
    :undoc-members:
@@ -93,8 +79,6 @@ It is also possible to use a table from a MySQL database to add fields to an eve
 """
 # pylint: enable=anomalous-backslash-in-string
 
-import re
-from typing import Any
 
 from attrs import define, field, validators
 
@@ -143,44 +127,6 @@ class GenericAdderRule(FieldManagerRule):
         """If a list is used, it is possible to tell the generic adder to only use the
         first existing file by setting :code:`generic_adder.only_first_existing_file: true`.
         In that case, only one file must exist."""
-        sql_table: dict = field(
-            validator=[
-                validators.instance_of(dict),
-                validators.deep_mapping(
-                    key_validator=validators.in_(
-                        ["pattern", "event_source_field", "destination_field_prefix"]
-                    ),
-                    value_validator=validators.instance_of(str),
-                ),
-            ],
-            factory=dict,
-        )
-        # pylint: disable=anomalous-backslash-in-string
-        """ sql config for generic adder (Optional)
-        If a specified field in the table matches a condition, the remaining fields,
-        except for the ID field, will be added to the event.
-        The names of the new fields correspond to the column names in the MySQL table.
-        This is mutually exclusive with the addition from a list.
-
-        It can be defined via :code:`generic_adder.sql_table`.
-        There :code:`generic_adder.sql_table.event_source_field` defines a field in the event that
-        is being compared with values in the column of the MySQL table defined
-        in the processor config. However, only a part of :code:`event_source_field` will
-        be compared.
-        Which part this is can be configured via :code:`generic_adder.sql_table.pattern`.
-        This is a regex pattern with a capture group.
-        The value in the capture group is being extracted and used for the comparison.
-        :code:`generic_adder.sql_table.destination_field_prefix` can be used to prefix all added
-        fields with a dotted path, creating a nested dictionary.
-
-        In the following example the value of the field :code:`source` is being parsed
-        with :code:`pattern: ([a-zA-Z0-9]+)_\S+`.
-        It extracts the first alphanumerical string delimited by :code:`_`.
-        I.e., :code:`Test0_foobarbaz` would extract :code:`test0`, which would be
-        used for the comparison in the MySQL table.
-        Since :code:`destination_field_prefix: nested.dict` is set,
-        a newly added field :code:`FOO_NEW` would be placed under :code:`nested.dict.FOO_NEW`.
-        """
 
         # pylint: enable=anomalous-backslash-in-string
 
@@ -217,19 +163,3 @@ class GenericAdderRule(FieldManagerRule):
     def add(self) -> dict:
         """Returns the fields to add"""
         return self._config.add
-
-    @property
-    def db_target(self) -> str:
-        """Returns the db target"""
-        return self._config.sql_table.get("event_source_field")
-
-    @property
-    def db_pattern(self) -> Any:
-        """Returns the db pattern"""
-        raw_db_pattern = self._config.sql_table.get("pattern")
-        return re.compile(raw_db_pattern) if raw_db_pattern else None
-
-    @property
-    def db_destination_prefix(self) -> str:
-        """Returns the destination prefix"""
-        return self._config.sql_table.get("destination_field_prefix", "")