+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * fix `delimiter` typo in `StringSplitterRule` configuration
 * removed the configuration `tld_lists` in `domain_resolver`, `domain_label_extractor` and `pseudonymizer` as
 the list is now fixed inside the packaged logprep
+* remove SQL feature from `generic_adder`, fields can only be added from rule config or from file
 
 ### Features